+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-* Create `BakeExampleProblemTitle` direction for `statistics`
+* Add optional baking exercise problem title to `BakeExample` direction for `statistics`
 * Add support for baking learning objectives in appendices (`BakeLearningObjectives`) for `marketing`
 
 ## [v1.8.0] - 2022-23-06